sủa đơn vị khố lượng sản phẩm sau đống gói

+++ src/main/java/org/example/ecommerce/entity/Inventory.java
@@ -66,7 +66,7 @@ public class Inventory {
     private Integer weight;
 
     @Column(name = "length")
-    private Integer length;
+    private Float length;
 
     @Column(name = "width")
     private Integer width;
